"pagination" added as a "common" component
