+++ src/features/movieList/MovieList.jsx
@@ -16,17 +16,12 @@ import {
   VoteAverage,
   VoteInfo,
   Year,
-  PaginationWrapper,
-  PaginationButton,
-  PageInfo,
-  Page,
-  ButtonParagraph,
   PosterPlaceholder,
   StyledVideoIcon,
 } from "./styled";
 import StarIcon from "./images/star.svg";
 import VideoIcon from "./images/video.svg";
-import { StyledVectorLeft, StyledVectorRight } from "./styled";
+import Pagination from "../../common/Pagination/Pagination";
 
 const IMG_BASE_URL = "https://image.tmdb.org/t/p/w300";
 
@@ -73,13 +68,6 @@ const MovieList = () => {
 
     fetchData();
   }, [currentPage]);
-
-  const goToFirstPage = () => setCurrentPage(1);
-  const goToPreviousPage = () =>
-    setCurrentPage((prev) => Math.max(prev - 1, 1));
-  const goToNextPage = () =>
-    setCurrentPage((prev) => Math.min(prev + 1, totalPages));
-  const goToLastPage = () => setCurrentPage(totalPages);
 
   if (loading) return <p>Loading...</p>;
   if (error) return <p>Error: {error}</p>;
@@ -140,48 +128,12 @@ const MovieList = () => {
           )
         )}
       </List>
-      <PaginationWrapper>
-        <PaginationButton onClick={goToFirstPage} disabled={currentPage === 1}>
-          <span className="double-arrow">
-            <StyledVectorLeft disabled={currentPage === 1} />
-            <StyledVectorLeft
-              className="mobile-only"
-              disabled={currentPage === 1}
-            />
-          </span>
-          <ButtonParagraph>First</ButtonParagraph>
-        </PaginationButton>
-        <PaginationButton
-          onClick={goToPreviousPage}
-          disabled={currentPage === 1}
-        >
-          <StyledVectorLeft disabled={currentPage === 1} />
-          <ButtonParagraph>Previous</ButtonParagraph>
-        </PaginationButton>
-        <PageInfo>
-          Page <Page>{currentPage}</Page> of <Page>{totalPages}</Page>
-        </PageInfo>
-        <PaginationButton
-          onClick={goToNextPage}
-          disabled={currentPage === totalPages}
-        >
-          <ButtonParagraph>Next</ButtonParagraph>{" "}
-          <StyledVectorRight disabled={currentPage === totalPages} />
-        </PaginationButton>
-        <PaginationButton
-          onClick={goToLastPage}
-          disabled={currentPage === totalPages}
-        >
-          <ButtonParagraph>Last</ButtonParagraph>
-          <span className="double-arrow">
-            <StyledVectorRight disabled={currentPage === totalPages} />
-            <StyledVectorRight
-              className="mobile-only"
-              disabled={currentPage === totalPages}
-            />
-          </span>
-        </PaginationButton>
-      </PaginationWrapper>
+
+      <Pagination
+        currentPage={currentPage}
+        totalPages={totalPages}
+        setCurrentPage={setCurrentPage}
+      />
     </Container>
   );
 };
