+++ src/features/movieList/styled.js
@@ -1,6 +1,4 @@
 import styled from "styled-components";
-import { ReactComponent as VectorRightIcon } from "./images/vectorRight.svg?react";
-import { ReactComponent as VectorLeftIcon } from "./images/vectorLeft.svg?react";
 
 export const Container = styled.div`
   max-width: 1368px;
@@ -147,109 +145,4 @@ export const VoteInfo = styled.span`
   font-size: 16px;
   line-height: 150%;
   color: #7e839a;
-`;
-
-export const PaginationWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-  margin-top: 40px;
-  font-size: 14px;
-  padding-bottom: 103px;
-`;
-
-export const ButtonParagraph = styled.p`
-  color: black;
-
-  @media (max-width: 667px) {
-    display: none;
-  }
-`;
-
-export const PaginationButton = styled.button`
-  background: #d6e4ff;
-  border: none;
-  border-radius: 5px;
-  padding: 6px 12px;
-  cursor: pointer;
-  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.1);
-  transition: background-color 0.3s ease;
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-  min-width: 75px;
-  height: 36px;
-  color: #0044cc;
-
-  &:hover:not(:disabled) {
-    background: #e0e0e0;
-  }
-
-  &:disabled {
-    opacity: 0.5;
-    cursor: default;
-    color: #7e839a;
-  }
-
-  .mobile-only {
-    display: none;
-  }
-
-  .double-arrow {
-    display: inline-flex;
-    align-items: center;
-    gap: 0px;
-  }
-
-  @media (max-width: 667px) {
-    min-width: 50px;
-    padding: 4px 8px;
-    gap: 4px;
-
-    .mobile-only {
-      display: inline;
-    }
-  }
-`;
-
-export const StyledVectorRight = styled(VectorRightIcon)`
-  width: 16px;
-  height: 16px;
-
-  & path {
-    fill: ${({ disabled }) => (disabled ? "#7e839a" : "#0044cc")};
-  }
-
-  @media (max-width: 667px) {
-    width: 10px;
-    height: 10px;
-  }
-`;
-
-export const StyledVectorLeft = styled(VectorLeftIcon)`
-  width: 16px;
-  height: 16px;
-
-  & path {
-    fill: ${({ disabled }) => (disabled ? "#7e839a" : "#0044cc")};
-  }
-
-  @media (max-width: 667px) {
-    width: 10px;
-    height: 10px;
-  }
-`;
-
-export const PageInfo = styled.span`
-  color: #7e839a;
-  font-size: 16px;
-`;
-
-export const Page = styled.span`
-  color: black;
-  font-weight: 600;
-  font-size: 16px;
-  line-height: 150%;
 `;
